Use unsigned int instead of size_t
git-svn-id: svn+ssh://subversion.munts.net/src/trunk/ARM@1607 06ee2f96-5407-11de-8bab-f3d7589a4122

+++ str912faw44/serial.c
@@ -234,7 +234,7 @@ static void serial_putch(unsigned port, char c)
 
 /* Send a buffer to the serial port */
 
-int serial_write(unsigned port, char *buf, size_t count)
+int serial_write(unsigned port, char *buf, unsigned int count)
 {
   int n;
 
@@ -253,7 +253,7 @@ int serial_rxready(unsigned port)
 
 /* Read buffer from the serial port */
 
-int serial_read(unsigned port, char *buf, size_t count)
+int serial_read(unsigned port, char *buf, unsigned int count)
 {
   if (serial_rxready(port))
   {
