+++ str912faw44/usb_serial/usb_serial.c
@@ -170,7 +170,7 @@ int usb_serial_rxready(unsigned subdevice)
 
 // Send data to USB host
 
-int usb_serial_write(unsigned subdevice, char *buf, size_t count)
+int usb_serial_write(unsigned subdevice, char *buf, unsigned int count)
 {
   if (!usb_serial_txready(subdevice)) return 0;
 
@@ -187,7 +187,7 @@ int usb_serial_write(unsigned subdevice, char *buf, size_t count)
 
 // Receive data from USB host
 
-int usb_serial_read(unsigned subdevice, char *buf, size_t count)
+int usb_serial_read(unsigned subdevice, char *buf, unsigned int count)
 {
   return ringbuffer_dequeue((ringbuffer_t *) &rxbuf, buf, count);
 }
